fixed import loop
