+++ python/semantic_kernel/connectors/ai/function_call_choice_configuration.py
@@ -1,0 +1,15 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+
+from pydantic.dataclasses import dataclass
+
+from semantic_kernel.functions.kernel_function_metadata import KernelFunctionMetadata
+from semantic_kernel.utils.experimental_decorator import experimental_class
+
+
+@experimental_class
+@dataclass
+class FunctionCallChoiceConfiguration:
+    """Configuration for function call choice."""
+
+    available_functions: list[KernelFunctionMetadata] | None = None
