+++ python/semantic_kernel/connectors/ai/function_calling_utils.py
@@ -1,15 +1,19 @@
 # Copyright (c) Microsoft. All rights reserved.
 
-from typing import Any
+from collections import OrderedDict
+from typing import TYPE_CHECKING, Any
 
-from semantic_kernel.connectors.ai.function_choice_behavior import FunctionCallChoiceConfiguration
-from semantic_kernel.connectors.ai.prompt_execution_settings import PromptExecutionSettings
-from semantic_kernel.functions.kernel_function_metadata import KernelFunctionMetadata
+from semantic_kernel.exceptions.service_exceptions import ServiceInitializationError
+
+if TYPE_CHECKING:
+    from semantic_kernel.connectors.ai.function_choice_behavior import FunctionCallChoiceConfiguration
+    from semantic_kernel.connectors.ai.prompt_execution_settings import PromptExecutionSettings
+    from semantic_kernel.functions.kernel_function_metadata import KernelFunctionMetadata
 
 
 def update_settings_from_function_call_configuration(
-    function_choice_configuration: FunctionCallChoiceConfiguration,
-    settings: PromptExecutionSettings,
+    function_choice_configuration: "FunctionCallChoiceConfiguration",
+    settings: "PromptExecutionSettings",
     type: str,
 ) -> None:
     """Update the settings from a FunctionChoiceConfiguration."""
@@ -26,7 +30,7 @@ def update_settings_from_function_call_configuration(
 
 
 def kernel_function_metadata_to_function_call_format(
-    metadata: KernelFunctionMetadata,
+    metadata: "KernelFunctionMetadata",
 ) -> dict[str, Any]:
     """Convert the kernel function metadata to function calling format."""
     return {
@@ -41,3 +45,26 @@ def kernel_function_metadata_to_function_call_format(
             },
         },
     }
+
+
+def _combine_filter_dicts(*dicts: dict[str, list[str]]) -> dict:
+    """Combine multiple filter dictionaries with list values into one dictionary.
+
+    This method is ensuring unique values while preserving order.
+    """
+    combined_filters = {}
+
+    keys = set().union(*(d.keys() for d in dicts))
+
+    for key in keys:
+        combined_functions: OrderedDict[str, None] = OrderedDict()
+        for d in dicts:
+            if key in d:
+                if isinstance(d[key], list):
+                    for item in d[key]:
+                        combined_functions[item] = None
+                else:
+                    raise ServiceInitializationError(f"Values for filter key '{key}' are not lists.")
+        combined_filters[key] = list(combined_functions.keys())
+
+    return combined_filters
